+++ arista/util/experimental/showtech_analyzer/changelog.md
@@ -1,3 +1,14 @@
+# v1.5.0
+- [BUG]: The i2c parser has been updated to handle STDERR cases (no -b -w was given) and defaults to Byte sized outputs.
+- [STYLE]: Refactored Platform Config detection to the file upload level from the system summary level.
+
+- [FEATURE]: I2C Support
+    - Added I2C devices field to Platform Configs
+    - Created register map for each of (UCD90320, ISL68226, PMBUS)
+    - Each call to `i2cdump` is parsed and matched with a device and register map
+    - If the device was not recognized or the register map was not found, do not parse, only `raw`
+
+
 # v1.4.1
 - Extract the hostname and add as Metadata to the JSON output.